feat: use seperated file for c code
